Feat/#93:TIL, 공지사항 멤버 권한 검증 추가 (#98)
* Rebase/#86:충돌해결
- 문제양식 활성화 기능 구현
- API 추가
- 활성화 양식 추가됨으로써 기존 기능 수정

* Rebase/#86:충돌해결
- 연관관계 수정

* Refactor/#86:오류 개선

* Feat/#93:공지사항 권한 검증 추가

* Feat/#93:TIL 권한 검증 추가

* Refactor/#93:오류 수정

* Doc/#1:템플릿 수정

* Refactor/#93:오류 수정

* Refactor/#93:오류 수정

+++ cluvr-api/src/main/java/com/example/cluvrapi/domain/notice/service/NoticeServiceImpl.java
@@ -8,7 +8,10 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.example.cluvrapi.domain.club.entity.Club;
 import com.example.cluvrapi.domain.club.repository.ClubRepository;
+import com.example.cluvrapi.domain.clubMember.entity.ClubMember;
+import com.example.cluvrapi.domain.clubMember.repository.ClubMemberRepository;
 import com.example.cluvrapi.domain.common.dto.PageResponseDto;
+import com.example.cluvrapi.domain.common.validator.ClubValidator;
 import com.example.cluvrapi.domain.notice.dto.reqeust.CreateNoticeRequestDto;
 import com.example.cluvrapi.domain.notice.dto.reqeust.UpdateNoticeRequestDto;
 import com.example.cluvrapi.domain.notice.dto.response.CreateNoticeResponseDto;
@@ -17,6 +20,8 @@ import com.example.cluvrapi.domain.notice.entity.Notice;
 import com.example.cluvrapi.domain.notice.repository.NoticeRepository;
 import com.example.cluvrapi.domain.user.entity.User;
 import com.example.cluvrapi.domain.user.repository.UserRepository;
+import com.example.cluvrapi.global.exception.BusinessException;
+import com.example.cluvrapi.global.response.ResponseCode;
 
 @Service
 @RequiredArgsConstructor
@@ -25,6 +30,8 @@ public class NoticeServiceImpl implements NoticeService {
 	private final UserRepository userRepository;
 	private final ClubRepository clubRepository;
 	private final NoticeRepository noticeRepository;
+	private final ClubMemberRepository clubMemberRepository;
+	private final ClubValidator clubValidator;
 
 	@Override
 	@Transactional
@@ -33,6 +40,11 @@ public class NoticeServiceImpl implements NoticeService {
 		User findUser = userRepository.findByIdOrElseThrow(userId);
 
 		Club findClub = clubRepository.findByIdOrElseThrow(clubId);
+
+		ClubMember findClubMember = clubMemberRepository.findByClubIdAndUserId(clubId, userId).orElseThrow(
+			() -> new BusinessException(ResponseCode.INVALID_REQUEST, "해당하는 멤버가 존재하지 않습니다.")
+		);
+		clubValidator.validateOwnerAndAdminRole(findClubMember.getClubMemberRole());
 
 		Notice notice = new Notice(findUser, findClub, createNoticeRequestDto.getTitle(),
 			createNoticeRequestDto.getContent());
@@ -44,7 +56,9 @@ public class NoticeServiceImpl implements NoticeService {
 
 	@Override
 	public InfoNoticeResponseDto findNoticeById(Long clubId, Long noticeId) {
-		return noticeRepository.findNoticeById(clubId, noticeId);
+		return noticeRepository.findNoticeById(clubId, noticeId).orElseThrow(
+			() -> new BusinessException(ResponseCode.NOT_FOUND, "존재하지 않는 공지사항 입니다.")
+		);
 	}
 
 	@Override
@@ -54,8 +68,13 @@ public class NoticeServiceImpl implements NoticeService {
 
 	@Override
 	@Transactional
-	public void updateNotice(Long noticeId, UpdateNoticeRequestDto updateNoticeRequestDto) {
+	public void updateNotice(Long userId, Long clubId, Long noticeId, UpdateNoticeRequestDto updateNoticeRequestDto) {
 		Notice findNotice = noticeRepository.findByIdOrElseThrow(noticeId);
+
+		ClubMember findClubMember = clubMemberRepository.findByClubIdAndUserId(clubId, userId).orElseThrow(
+			() -> new BusinessException(ResponseCode.INVALID_REQUEST, "해당하는 멤버가 존재하지 않습니다.")
+		);
+		clubValidator.validateOwnerAndAdminRole(findClubMember.getClubMemberRole());
 
 		if (updateNoticeRequestDto.getTitle() != null) {
 			findNotice.updateTitle(updateNoticeRequestDto.getTitle());
@@ -68,7 +87,12 @@ public class NoticeServiceImpl implements NoticeService {
 
 	@Override
 	@Transactional
-	public void deleteNotice(Long noticeId) {
+	public void deleteNotice(Long userId, Long clubId, Long noticeId) {
+		ClubMember findClubMember = clubMemberRepository.findByClubIdAndUserId(clubId, userId).orElseThrow(
+			() -> new BusinessException(ResponseCode.INVALID_REQUEST, "해당하는 멤버가 존재하지 않습니다.")
+		);
+		clubValidator.validateOwnerAndAdminRole(findClubMember.getClubMemberRole());
+
 		Notice findNotice = noticeRepository.findByIdOrElseThrow(noticeId);
 
 		noticeRepository.delete(findNotice);
